associate todo with user, logout user, @login_require decorator, custom decorator to hide login and register when authenticated

+++ todo/views.py
@@ -4,6 +4,7 @@ from django.shortcuts import get_object_or_404, render
 from django.http.response import HttpResponseRedirect
 from django.urls import reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 
 def get_showing_todos(request, todos):
@@ -16,8 +17,10 @@ def get_showing_todos(request, todos):
     return todos
 
 
+@login_required
 def index(request):
-    todos = Todo.objects.all()
+
+    todos = Todo.objects.filter(owner=request.user)
 
     completed_count = todos.filter(is_completed=True).count()
     incomplete_count = todos.filter(is_completed=False).count()
@@ -31,6 +34,7 @@ def index(request):
     return render(request, 'todo/index.html', context)
 
 
+@login_required
 def create_todo(request):
     form = TodoForm()
     context = {'form': form}
@@ -44,6 +48,7 @@ def create_todo(request):
         todo.title = title
         todo.description = description
         todo.is_completed = True if is_completed == "on" else False
+        todo.owner = request.user
 
         todo.save()
 
@@ -55,26 +60,35 @@ def create_todo(request):
     return render(request, 'todo/create-todo.html', context)
 
 
+@login_required
 def todo_detail(request, id):
-    todo = get_object_or_404(Todo, pk=id)
+    todo = get_object_or_404(Todo, owner=request.user, pk=id)
     context = {'todo': todo}
     return render(request, 'todo/todo-detail.html', context)
 
 
+@login_required
 def todo_delete(request, id):
     todo = get_object_or_404(Todo, pk=id)
     context = {'todo': todo}
 
     if request.method == 'POST':
-        todo.delete()
 
-        messages.add_message(request, messages.SUCCESS, "Todo delete success")
+        if todo.owner == request.user:
 
-        return HttpResponseRedirect(reverse('home44'))
+            todo.delete()
+
+            messages.add_message(request, messages.SUCCESS,
+                                 "Todo delete success")
+
+            return HttpResponseRedirect(reverse('home44'))
+
+        return render(request, 'todo/todo-delete.html', context)
 
     return render(request, 'todo/todo-delete.html', context)
 
 
+@login_required
 def todo_edit(request, id):
     todo = get_object_or_404(Todo, pk=id)
     form = TodoForm(instance=todo)
@@ -87,10 +101,12 @@ def todo_edit(request, id):
         todo.is_completed = True if request.POST.get(
             'is_completed', False) == "on" else False
 
-        todo.save()
+        if todo.owner == request.user:
+            todo.save()
 
-        messages.add_message(request, messages.SUCCESS, "Todo update success")
+            messages.add_message(request, messages.SUCCESS,
+                                 "Todo update success")
 
-        return HttpResponseRedirect(reverse('todo66', kwargs={'id': todo.pk}))
+            return HttpResponseRedirect(reverse('todo66', kwargs={'id': todo.pk}))
 
     return render(request, 'todo/todo-edit.html', context)
